pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integra"
version = "0.1.0"
description = "A multi-agent system for integrating open-source solutions into PydanticAI projects"
readme = "README.md"
authors = [
    {name = "Integra Team"}
]
requires-python = ">=3.9"
dependencies = [
    "pydantic",
    "pydantic_ai",
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.10.0",
    "markdown>=3.4.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "duckduckgo-search>=4.1.1",
    "PyGithub>=2.1.1",
    "pypi-simple>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
] 